refactors the tests into separate files
