+++ tests/test_generate.py
@@ -1,0 +1,15 @@
+from unittest import TestCase
+
+# from nhs_number import generate, is_valid
+
+"""
+* [ ] test for a valid number
+* [ ] test for invalid number generation
+* [ ] test for specific country validity
+
+"""
+
+# class GenerationTests(TestCase):
+
+#     def test_generate_random_valid_number(self):
+#     generate()
